# perf(es/parser): Improve performance