--- a/crates/swc_ecma_loader/src/resolvers/node.rs
+++ b/crates/swc_ecma_loader/src/resolvers/node.rs
@@ -125,11 +125,11 @@ impl NodeModulesResolver {
                 let extensions = match old_ext {
                     // Note that the official compiler code always tries ".ts" before
                     // ".tsx" even if the original extension was ".jsx".
-                    "js" => ["ts", "tsx"].as_slice(),
-                    "jsx" => ["ts", "tsx"].as_slice(),
-                    "mjs" => ["mts"].as_slice(),
-                    "cjs" => ["cts"].as_slice(),
-                    _ => [].as_slice(),
+                    "js" => &["ts", "tsx"],
+                    "jsx" => &["ts", "tsx"],
+                    "mjs" => &["mts"],
+                    "cjs" => &["cts"],
+                    _ => &[],
                 };
 
                 for ext in extensions {