--- a/crates/swc_ecma_loader/src/resolvers/node.rs
+++ b/crates/swc_ecma_loader/src/resolvers/node.rs
@@ -122,13 +122,13 @@ impl NodeModulesResolver {
             let old_ext = path.extension().and_then(|ext| ext.to_str());
 
             if let Some(old_ext) = old_ext {
-                let extensions = match old_ext {
+                let extensions: &[&str] = match old_ext {
                     // Note that the official compiler code always tries ".ts" before
                     // ".tsx" even if the original extension was ".jsx".
-                    "js" => &*["ts", "tsx"],
-                    "jsx" => &*["ts", "tsx"],
-                    "mjs" => &*["mts"],
-                    "cjs" => &*["cts"],
+                    "js" => &["ts", "tsx"],
+                    "jsx" => &["ts", "tsx"],
+                    "mjs" => &["mts"],
+                    "cjs" => &["cts"],
                     _ => &[],
                 };