--- a/crates/swc_ecma_parser/src/parser/class_and_fn.rs
+++ b/crates/swc_ecma_parser/src/parser/class_and_fn.rs
@@ -455,7 +455,6 @@ impl<'a, I: Tokens> Parser<I> {
         Ok(StaticBlock { span, body }.into())
     }
 
-    #[allow(clippy::cognitive_complexity)]
     fn parse_class_member_with_is_static(
         &mut self,
         start: BytePos,