--- a/crates/swc_ecma_parser/src/parser/class_and_fn.rs
+++ b/crates/swc_ecma_parser/src/parser/class_and_fn.rs
@@ -607,19 +607,12 @@ impl<'a, I: Tokens> Parser<I> {
         };
         let is_optional = self.input.syntax().typescript() && eat!(self, '?');
 
-        let is_private = match key {
-            Either::Left(PrivateName { .. }) => true,
-            _ => false,
-        };
-        let is_simple = {
-            match &mut key {
-                Either::Right(PropName::Ident(i)) => {
-                    i.optional = is_optional;
-                    true
-                }
-                _ => false,
+        match &mut key {
+            Either::Right(PropName::Ident(i)) => {
+                i.optional = is_optional;
             }
-        };
+            _ => {}
+        }
 
         if self.is_class_method() {
             // handle a(){} / get(){} / set(){} / async(){}
@@ -996,7 +989,6 @@ impl<'a, I: Tokens> Parser<I> {
         let is_async = start_of_async.is_some();
 
         let is_generator = {
-            let start = cur_pos!(self);
             if eat!(self, '*') {
                 // if is_async {
                 //     syntax_error!(self, span!(self, start), SyntaxError::AsyncGenerator {});
@@ -1024,7 +1016,6 @@ impl<'a, I: Tokens> Parser<I> {
             // function declaration does not change context for `BindingIdentifier`.
             self.parse_maybe_opt_binding_ident()?
         };
-        let is_constructor = T::is_constructor(&ident);
 
         self.parse_with(|p| {
             let f = p.parse_fn_args_body(
@@ -1373,7 +1364,7 @@ impl OutputType for Decl {
             function,
         })
     }
-    fn finish_class(span: Span, ident: Ident, class: Class) -> Self {
+    fn finish_class(_: Span, ident: Ident, class: Class) -> Self {
         Decl::Class(ClassDecl {
             declare: false,
             ident,
@@ -1449,6 +1440,8 @@ struct MakeMethodArgs {
 
 #[cfg(test)]
 mod tests {
+    #![allow(unused)]
+
     use super::*;
     use swc_common::DUMMY_SP as span;
     use swc_ecma_visit::assert_eq_ignore_span;