--- a/crates/swc_ecma_parser/src/parser/expr.rs
+++ b/crates/swc_ecma_parser/src/parser/expr.rs
@@ -2,7 +2,7 @@ use super::{pat::PatType, util::ExprExt, *};
 use crate::{lexer::TokenContext, token::AssignOpToken};
 use either::Either;
 use swc_atoms::js_word;
-use swc_common::{ast_node, Spanned};
+use swc_common::{ast_node, util::take::Take, Spanned};
 
 mod ops;
 #[cfg(test)]
@@ -978,8 +978,11 @@ impl<'a, I: Tokens> Parser<I> {
     }
 
     /// returned bool is true if this method should be called again.
-    #[allow(clippy::cognitive_complexity)]
-    fn parse_subscript(&mut self, obj: ExprOrSuper, no_call: bool) -> PResult<(Box<Expr>, bool)> {
+    fn parse_subscript(
+        &mut self,
+        mut obj: ExprOrSuper,
+        no_call: bool,
+    ) -> PResult<(Box<Expr>, bool)> {
         let _ = cur!(self, false);
         let start = obj.span().lo();
 
@@ -1009,15 +1012,20 @@ impl<'a, I: Tokens> Parser<I> {
                 }
             } && is!(self, '<')
             {
-                let obj_ref = &obj;
+                let is_dynamic_import = is_import(&obj);
+
+                let mut obj_opt = Some(obj);
                 // tsTryParseAndCatch is expensive, so avoid if not necessary.
                 // There are number of things we are going to "maybe" parse, like type arguments
                 // on tagged template expressions. If any of them fail, walk it back and
                 // continue.
+
+                let mut_obj_opt = &mut obj_opt;
+
                 let result = self.try_parse_ts(|p| {
                     if !no_call
-                        && p.at_possible_async(match obj_ref {
-                            ExprOrSuper::Expr(ref expr) => &*expr,
+                        && p.at_possible_async(match &mut_obj_opt {
+                            Some(ExprOrSuper::Expr(ref expr)) => &**expr,
                             _ => unreachable!(),
                         })?
                     {
@@ -1034,12 +1042,12 @@ impl<'a, I: Tokens> Parser<I> {
                     if !no_call && is!(p, '(') {
                         // possibleAsync always false here, because we would have handled it
                         // above. (won't be any undefined arguments)
-                        let args = p.parse_args(is_import(&obj))?;
+                        let args = p.parse_args(is_dynamic_import)?;
 
                         Ok(Some((
                             Box::new(Expr::Call(CallExpr {
                                 span: span!(p, start),
-                                callee: obj_ref.clone(),
+                                callee: mut_obj_opt.take().unwrap(),
                                 type_args: Some(type_args),
                                 args,
                             })),
@@ -1047,8 +1055,8 @@ impl<'a, I: Tokens> Parser<I> {
                         )))
                     } else if is!(p, '`') {
                         p.parse_tagged_tpl(
-                            match *obj_ref {
-                                ExprOrSuper::Expr(ref obj) => obj.clone(),
+                            match mut_obj_opt {
+                                Some(ExprOrSuper::Expr(obj)) => obj.take(),
                                 _ => unreachable!(),
                             },
                             Some(type_args),
@@ -1066,6 +1074,8 @@ impl<'a, I: Tokens> Parser<I> {
                 if let Some(result) = result {
                     return Ok(result);
                 }
+
+                obj = obj_opt.unwrap();
             }
         }