--- a/crates/swc_ecma_parser/src/parser/stmt.rs
+++ b/crates/swc_ecma_parser/src/parser/stmt.rs
@@ -993,7 +993,6 @@ impl<'a, I: Tokens> Parser<I> {
     }
 
     fn parse_for_head(&mut self) -> PResult<ForHead> {
-        let start = cur_pos!(self);
         let strict = self.ctx().strict;
 
         if is_one_of!(self, "const", "var")
@@ -1002,8 +1001,6 @@ impl<'a, I: Tokens> Parser<I> {
             let decl = self.parse_var_stmt(true)?;
 
             if is_one_of!(self, "of", "in") {
-                let is_in = is!(self, "in");
-
                 if decl.decls.len() != 1 {
                     for d in decl.decls.iter().skip(1) {
                         self.emit_err(d.name.span(), SyntaxError::TooManyVarInForInHead);
@@ -1056,7 +1053,7 @@ impl<'a, I: Tokens> Parser<I> {
             // for ({} in foo) is invalid
             if self.input.syntax().typescript() && is_in {
                 match pat {
-                    Pat::Ident(ref v) => {}
+                    Pat::Ident(..) => {}
                     Pat::Expr(..) => {}
                     ref v => self.emit_err(v.span(), SyntaxError::TS2491),
                 }
@@ -1172,7 +1169,7 @@ where
 }
 
 impl<'a, I: Tokens> StmtLikeParser<'a, Stmt> for Parser<I> {
-    fn handle_import_export(&mut self, top_level: bool, _: Vec<Decorator>) -> PResult<Stmt> {
+    fn handle_import_export(&mut self, _: bool, _: Vec<Decorator>) -> PResult<Stmt> {
         let start = cur_pos!(self);
         if self.input.syntax().dynamic_import() && is!(self, "import") && peeked_is!(self, '(') {
             let expr = self.parse_expr()?;