--- a/crates/swc_ecma_parser/src/parser/typescript.rs
+++ b/crates/swc_ecma_parser/src/parser/typescript.rs
@@ -2635,8 +2635,13 @@ impl<I: Tokens> Parser<I> {
         debug_assert!(self.input.syntax().typescript());
 
         let cloned = self.input.token_context().clone();
+
+        #[cfg(debug_assertions)]
         self.input
             .set_token_context(TokenContexts(vec![cloned.0[0]]));
+        #[cfg(not(debug_assertions))]
+        self.input
+            .set_token_context(TokenContexts(smallvec::smallvec![cloned.0[0]]));
         let res = op(self);
         self.input.set_token_context(cloned);