swc_common:
 - Optimize `StringInput`.

swc_ecma_parser:
 - Optimize `skip_space`.
 - Remove useless `.clone()`.



---

### main

```
test full_es5                     ... bench: 112,497,693 ns/iter (+/- 5,207,775)
```

```
test angular       ... bench:  12,603,811 ns/iter (+/- 4,229,331) = 57 MB/s
test backbone      ... bench:   1,726,430 ns/iter (+/- 445,682) = 34 MB/s
test colors        ... bench:      25,007 ns/iter (+/- 2,100) = 46 MB/s
test jquery        ... bench:   9,356,590 ns/iter (+/- 2,128,406) = 28 MB/s
test jquery_mobile ... bench:  14,626,101 ns/iter (+/- 2,907,195) = 30 MB/s
test mootools      ... bench:   7,249,361 ns/iter (+/- 1,981,647) = 22 MB/s
test underscore    ... bench:   1,501,914 ns/iter (+/- 122,972) = 29 MB/s
test yui           ... bench:   7,125,644 ns/iter (+/- 1,480,380) = 47 MB/s
```


### After 3f2fb01339b3599faf3a9f52dd3fc0e51f5328d4

```
test full_es5                     ... bench: 112,386,755 ns/iter (+/- 7,092,674)
```

### After 606f75958ddf53444bd298aed293a111627e1365

```
test full_es5                     ... bench: 112,013,618 ns/iter (+/- 5,518,629)
```


### After 4bd2517768c1ff5f330040711c4dff4e44322f1c

```
test full_es5                     ... bench: 110,520,237 ns/iter (+/- 5,287,594)
```



```
test angular       ... bench:  12,711,254 ns/iter (+/- 4,474,660) = 56 MB/s
test backbone      ... bench:   1,805,933 ns/iter (+/- 428,122) = 33 MB/s
test colors        ... bench:      24,310 ns/iter (+/- 1,626) = 47 MB/s
test jquery        ... bench:   9,193,212 ns/iter (+/- 1,967,762) = 29 MB/s
test jquery_mobile ... bench:  14,290,825 ns/iter (+/- 2,748,184) = 31 MB/s
test mootools      ... bench:   7,242,212 ns/iter (+/- 1,940,175) = 22 MB/s
test underscore    ... bench:   1,472,203 ns/iter (+/- 84,305) = 29 MB/s
test yui           ... bench:   6,939,852 ns/iter (+/- 2,089,068) = 48 MB/s
```


### After 48c399058b4c3f61faf1a3c37d4876d709ac1bfa


```
test angular       ... bench:  12,395,044 ns/iter (+/- 3,184,236) = 57 MB/s
test backbone      ... bench:   1,688,445 ns/iter (+/- 145,841) = 35 MB/s
test colors        ... bench:      24,114 ns/iter (+/- 833) = 48 MB/s
test jquery        ... bench:   9,144,634 ns/iter (+/- 1,967,653) = 29 MB/s
test jquery_mobile ... bench:  14,494,202 ns/iter (+/- 3,277,632) = 31 MB/s
test mootools      ... bench:   7,163,693 ns/iter (+/- 2,356,552) = 22 MB/s
test underscore    ... bench:   1,473,210 ns/iter (+/- 120,565) = 29 MB/s
test yui           ... bench:   7,048,669 ns/iter (+/- 1,790,961) = 47 MB/s
```

### After 3392bc2b77e8837f6270c87af4270d1dadbabfa4

```
test angular       ... bench:  12,367,571 ns/iter (+/- 2,983,180) = 58 MB/s
test backbone      ... bench:   1,703,499 ns/iter (+/- 212,168) = 35 MB/s
test colors        ... bench:      24,201 ns/iter (+/- 1,010) = 47 MB/s
test jquery        ... bench:   9,144,588 ns/iter (+/- 1,512,935) = 29 MB/s
test jquery_mobile ... bench:  14,397,090 ns/iter (+/- 3,271,368) = 31 MB/s
test mootools      ... bench:   7,113,694 ns/iter (+/- 1,763,498) = 22 MB/s
test underscore    ... bench:   1,480,600 ns/iter (+/- 96,628) = 29 MB/s
test yui           ... bench:   6,976,067 ns/iter (+/- 1,747,759) = 48 MB/s
```


### After 853f128b2a6a76ce924bc2559132d6d3dc53d95e

```
test angular       ... bench:  12,361,690 ns/iter (+/- 4,233,363) = 58 MB/s
test backbone      ... bench:   1,703,948 ns/iter (+/- 370,609) = 35 MB/s
test colors        ... bench:      24,413 ns/iter (+/- 1,075) = 47 MB/s
test jquery        ... bench:   9,271,220 ns/iter (+/- 1,683,387) = 28 MB/s
test jquery_mobile ... bench:  14,575,688 ns/iter (+/- 3,893,379) = 31 MB/s
test mootools      ... bench:   7,256,533 ns/iter (+/- 2,031,723) = 22 MB/s
test underscore    ... bench:   1,490,333 ns/iter (+/- 147,705) = 29 MB/s
test yui           ... bench:   6,970,418 ns/iter (+/- 1,549,417) = 48 MB/s
```